requirements1.txt
bitarray==3.7.2
cupy==13.6.0
faiss==1.7.3
GitPython==3.1.45
numpy==1.24.4
regex==2024.11.6
torch==1.13.1
tqdm==4.67.1
transformers==4.46.3
ujson==5.10.0
